chore: Remove unused imports and update file structure
