+++ lib/core/shared/shared.dart
@@ -12,7 +12,23 @@ export 'remote/network_info.dart';
 export 'remote/endpoints.dart';
 export 'remote/response.dart';
 export 'router.dart';
+export 'package:cached_network_image/cached_network_image.dart';
+export 'package:flutter/material.dart' hide ThemeExtension;
+export 'package:flutter_animate/flutter_animate.dart';
+export 'package:flutter_bloc/flutter_bloc.dart';
+export 'package:go_router/go_router.dart';
+export 'package:equatable/equatable.dart';
+export 'package:flutter_native_splash/flutter_native_splash.dart';
 
+export 'dart:async';
+export 'dart:io';
+export 'package:flutter/services.dart';
+
+export 'package:get_it/get_it.dart';
+export 'package:http/http.dart';
+export 'package:hydrated_bloc/hydrated_bloc.dart';
+export 'package:internet_connection_checker/internet_connection_checker.dart';
+export 'package:path_provider/path_provider.dart';
 
 export 'enums.dart';
 export 'text_styles.dart';
