+++ lib/core/shared/theme/theme_bloc.dart
@@ -1,5 +1,3 @@
-import 'package:equatable/equatable.dart';
-import 'package:hydrated_bloc/hydrated_bloc.dart';
 
 import '../shared.dart';
