+++ lib/features/analysis/data/datasources/remote_impl.dart
@@ -1,7 +1,4 @@
-import 'dart:io';
 
-import 'package:flutter/services.dart';
-import 'package:http/http.dart';
 
 import '../../../../core/shared/shared.dart';
 import '../../analysis.dart';
