+++ lib/features/fixture/fixture.dart
@@ -13,4 +13,5 @@ export 'data/datasources/local_impl.dart';
 export 'presentation/bloc/find_fixture_by_id_bloc.dart';
 export 'presentation/bloc/fixtures_bloc.dart';
 export 'presentation/pages/fixtures.dart';
+export 'presentation/widgets/fixtures/item.dart';
 export 'presentation/pages/details.dart';
