+++ lib/features/team/data/datasources/remote_impl.dart
@@ -1,8 +1,3 @@
-import 'dart:io';
-
-import 'package:flutter/services.dart';
-import 'package:http/http.dart';
-
 import '../../../../core/shared/shared.dart';
 import '../../team.dart';
