+++ lib/features/team/presentation/widgets/details.dart
@@ -1,8 +1,3 @@
-import 'package:cached_network_image/cached_network_image.dart';
-import 'package:flutter/material.dart';
-import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:podcast/features/team/presentation/bloc/team_bloc.dart';
-
 import '../../../../core/shared/shared.dart';
 import '../../team.dart';
