+++ lib/features/team/presentation/widgets/name.dart
@@ -1,7 +1,3 @@
-
-import 'package:flutter/material.dart';
-import 'package:flutter_bloc/flutter_bloc.dart';
-
 import '../../../../core/shared/shared.dart';
 import '../../team.dart';
