+++ lib/main.dart
@@ -1,7 +1,3 @@
-import 'package:flutter/material.dart';
-import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:flutter_native_splash/flutter_native_splash.dart';
-
 import 'core/config/config.dart';
 import 'core/shared/shared.dart';
