+++ mock/fixtures.json
@@ -4,25 +4,25 @@
     "result": [
         {
             "fixtureId": "fc42542f-38ba-4f02-8773-428e22fead9d",
-            "matchName": "2nd ODI",
-            "tournamentName": "Cycle Pure Agarbatti Series",
+            "matchTitle": "Bangladesh vs Sri Lanka 2nd ODI",
+            "matchDescription": "Bangladesh vs Sri Lanka 2nd ODI, live from Johur Ahmed Cricket Stadium. ",
             "startDate": "2024-05-02T15:00:00",
             "result": null,
             "homeTeamId": "c4eaacd5-6e17-44ee-a82b-9cef84f9f7c3",
             "awayTeamId": "b4566820-b923-4045-aac6-4970690ef101",
-            "stadiumName": "Eden garden kolkata",
-            "logo":"https://resize.indiatvnews.com/en/resize/newbucket/1200_-/2024/04/mi-csk-highlights-1713095803.jpg"
+            "stadiumName": "Johur Ahmed Cricket Stadium",
+            "logo": "https://resize.indiatvnews.com/en/resize/newbucket/1200_-/2024/04/mi-csk-highlights-1713095803.jpg"
         },
         {
             "fixtureId": "1e23471a-3027-4df8-8bcd-d80702caa1fe",
-            "matchName": "1st T20I",
-            "tournamentName": "Walton Series",
+            "matchTitle": "IPL - CSK vs MI Live now.",
+            "matchDescription": "IPL - Match day 41, CSK vs MI live from Wankhede Stadium.",
             "startDate": "2024-05-02T19:00:00",
             "result": null,
             "homeTeamId": "50375151-f2f0-43d6-a679-825481faf3c2",
             "awayTeamId": "a0b35646-f74c-47ff-81c0-d699dc3b9588",
-            "stadiumName": "Johur ahmed stadium.",
-            "logo":"https://static.toiimg.com/thumb/msid-108208018,width-320,height-320,resizemode-4,imgsize-107262/108208018.jpg"
+            "stadiumName": "Wankhede Stadium",
+            "logo": "https://static.toiimg.com/thumb/msid-108208018,width-320,height-320,resizemode-4,imgsize-107262/108208018.jpg"
         }
     ]
 }
